chore: clean up the footer folder and move manage preferences to cookies banner folder
