+++ src/components/CookiesBanner/ManagePreferences.js
